Improv: separate *possibly* common and sensor dedicated modules for skju_core
