+++ skju_core/src/common.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Clone, Copy)]
+pub struct Coord {
+    pub x: f32,
+    pub y: f32,
+}
